pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-showcase-causal-genomics-demo"
version = "0.1.0"
description = "Causal genomics ML demo: a clean, reproducible showcase pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ali Shakeri Abroudi" }]
dependencies = [
  "numpy",
  "scikit-learn",
  "pyyaml",
  "torch",
  "pandas",
  "tqdm",
  "matplotlib",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.5", "black>=24.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_showcase*", "src*"]
exclude = ["tests*", "configs*", "scripts*", "build*", "*.egg-info*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
